+++ Changelog.md
@@ -1,3 +1,9 @@
+# 4.10.1
+  - Fixed a bug where move hints would only select locations in your own world
+  - Progression items that are not unique can now be show as Wahay of the Duo/Archipelago if it's logcially hard-required
+  - Cryptic hint generation has been sped up
+  - Added proper logic to fruities (WW slot machines)
+
 # 4.10
   - New cryptic hints
     - Wahay of the Duo: A unique item that's required to beat your seed.